🔨 채팅 시간 수정 및 뷰 수정
- 안쓰는 컴포넌트 삭제 추가
