+++ discatch/src/components/Header.jsx
@@ -7,7 +7,7 @@ import { flexBox } from "../shared/style";
 
 // route
 import { Link } from "react-router-dom";
-import { useLocation } from 'react-router-dom';
+import { useLocation } from "react-router-dom";
 
 // element
 import { Grid } from "../elements";
@@ -26,41 +26,53 @@ const Header = (props) => {
   const dispatch = useDispatch();
   const path = props.path;
   const userInfo = useSelector((state) => state.mypage.userInfo);
+  console.log(locationA);
+  console.log(path);
+  console.log(userInfo);
   let location1;
   if (userInfo.locationList && path.length === 1) {
-    location1 = userInfo?.locationList[0].split('@')[0]
+    location1 = userInfo?.locationList[0].split("@")[0];
   } else {
     location1 = locationA.state?.userLocation;
   }
-  const locationList = userInfo.locationList?.filter(v => v.split('@')[0] !== location1);
+  const locationList = userInfo.locationList?.filter(
+    (v) => v.split("@")[0] !== location1
+  );
+
   let location2;
   let location3;
   if (userInfo.locationList !== undefined) {
-    location2 = locationList[0].split('@')[0];
-    location3 = locationList[1].split('@')[0];
+    location2 = locationList[0].split("@")[0];
+    location3 = locationList[1].split("@")[0];
   }
 
   const [searchModal, setSearchModal] = useState(false);
-  
+
   useEffect(() => {
     dispatch(mypageActions._getUserInfo());
   }, []);
-  
+
   useEffect(() => {
     dispatch(searchMap(locationA.state?.location));
   }, []);
 
   let options;
   if (location1 !== undefined && location2 === undefined) {
-    options = [
-      { key: 1, value: location1 }
-    ]
-  } else if (location1 !== undefined && location2 !== undefined && location3 === undefined) {
+    options = [{ key: 1, value: location1 }];
+  } else if (
+    location1 !== undefined &&
+    location2 !== undefined &&
+    location3 === undefined
+  ) {
     options = [
       { key: 1, value: location1 },
-      { key: 2, value: location2 }
-            ];
-  } else if (location1 !== undefined && location2 !== undefined && location3 !== undefined) {
+      { key: 2, value: location2 },
+    ];
+  } else if (
+    location1 !== undefined &&
+    location2 !== undefined &&
+    location3 !== undefined
+  ) {
     options = [
       { key: 1, value: location1 },
       { key: 2, value: location2 },
@@ -75,6 +87,9 @@ const Header = (props) => {
     dispatch(searchMap(keyword));
   };
 
+  if (!userInfo) {
+    return <div></div>;
+  }
   return (
     <HeaderStyle>
       <Grid
@@ -86,13 +101,14 @@ const Header = (props) => {
         }}
       >
         <Grid width="20%" height="100%" margin="auto">
-          {path === "/" || path === "/community" || path === "/map/:village"? (
+          {path === "/" || path === "/community" || path === "/map/:village" ? (
             <SelectStyle onChange={onChangeHandler} value={place}>
-              {options && options.map((pl, idx) => (
-                <option key={pl.key} value={pl.value}>
-                  {pl.value}
-                </option>
-              ))}
+              {options &&
+                options.map((pl, idx) => (
+                  <option key={pl.key} value={pl.value}>
+                    {pl.value}
+                  </option>
+                ))}
             </SelectStyle>
           ) : (
             <ArrowLeft
