--- discatch/src/components/chat/ChatTest.css
@@ -1,130 +1,0 @@
-/* * {
-  box-sizing: border-box;
-}
-
-button {
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-}
-
-body {
-  height: auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-} */
-
-.background {
-  width: 100%;
-  max-width: 1200px;
-  position: absolute;
-  top: 0;
-  left: 0;
-}
-
-.container {
-  position: relative;
-  width: 100%;
-  min-height: 540px;
-  background-color: #ffffff;
-  overflow: hidden;
-}
-
-.header {
-  display: flex;
-  height: 60px;
-  padding: 12px 5px 8px;
-  align-items: center;
-  box-shadow: 0 0 1px #999999;
-}
-
-.info-block {
-  margin-left: 12px;
-  margin-right: 95px;
-}
-
-.username {
-  margin: 0 0 3px;
-  font-size: 16px;
-}
-
-.status {
-  margin: 0;
-  font-size: 12px;
-  color: #999999;
-}
-
-.call-box {
-  display: flex;
-  width: 60px;
-  justify-content: space-between;
-}
-
-#chat-box {
-  height: 400px;
-  padding: 12px 10px;
-  overflow: auto;
-}
-
-.bubble-box {
-  display: flex;
-}
-
-.bubble {
-  margin: 5px 0;
-  display: flex;
-  max-width: 300px;
-  font-size: 14px;
-}
-
-.friend-bubble {
-  background-color: #fbd986;
-  border-radius: 14px 14px 14px 0;
-  padding: 7px 15px 7px 15px;
-  float: left;
-  clear: both;
-  color: #000000;
-}
-
-.my-bubble {
-  background-color: #cbcf52;
-  border-radius: 14px 14px 14px 0;
-  padding: 7px 15px 7px 15px;
-  float: right;
-  clear: both;
-  color: #000000;
-}
-
-.date {
-  font-size: 10px;
-  margin: 0px;
-}
-
-.footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  width: 100%;
-  height: 80px;
-  padding: 10px;
-  position: absolute;
-  bottom: 5;
-}
-
-#input {
-  width: 88%;
-  height: 100%;
-  padding: 10px;
-  background-color: #ffffff;
-  border: 1px solid #cbcf52;
-  border-radius: 10px;
-  resize: none;
-}
-
-.sendBtn {
-  width: 25px;
-  height: 25px;
-  margin-bottom: 10px;
-  stroke: #cbcf52;
-}
