+++ discatch/src/pages/chat/ChatRoomPage.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import { Template, ChatEx } from "../../components";
+import { Template, ChatRoom } from "../../components";
 const ChatRoomPage = (props) => {
   const roomId = props.match.params.roomId;
   console.log(roomId);
   return (
     <Template props={props} page="map">
-      <ChatEx roomId={roomId} />
+      <ChatRoom roomId={roomId} />
     </Template>
   );
 };
